+++ portofino-core/src/main/java/com/manydesigns/portofino/resourceactions/AbstractResourceAction.java
@@ -150,7 +150,7 @@ public abstract class AbstractResourceAction extends AbstractResourceWithParamet
         ActionContext context = new ActionContext();
         context.setRequest(request);
         context.setResponse(response);
-        context.setServletContext(request.getServletContext());
+        context.setServletContext(ElementsThreadLocals.getServletContext());
         if(uriInfo != null) { //TODO for Swagger
             String path = uriInfo.getPath();
             if (!path.startsWith("/")) {
@@ -327,8 +327,9 @@ public abstract class AbstractResourceAction extends AbstractResourceWithParamet
     }
 
     /**
-     * Returns an error response with message saying that the resourceaction is not properly
+     * Returns an error response with message saying that the resource-action is not properly
      * configured.
+     * @return the {@link Response}.
      */
     public Response resourceActionNotConfigured() {
         return Response.serverError().entity("resource-action-not-configured").build();