+++ portofino-core/src/main/java/com/manydesigns/portofino/spring/PortofinoSpringConfiguration.java
@@ -92,5 +92,4 @@ public class PortofinoSpringConfiguration {
         return new CacheResetListenerRegistry();
     }
 
-
 }