+++ ui/projects/portofino/src/lib/i18n/it.ts
@@ -70,6 +70,7 @@ export const TRANSLATIONS_IT = {
   "Old password": "Vecchia password",
   "Old password is required.": "Campo obbligatorio",
   "Page not found! Error: _": "Pagina non trovata! Errore: {{error}}",
+  "page type: custom": "Pagina personalizzata (custom)",
   "page type: crud": "Pagina CRUD",
   "page type: manyToMany": "Molti-a-molti",
   "page type: text": "Pagina di testo",