+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project from version 5.0.0 upwards are documented in this file. 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [5.2.1] – Unreleased
+
+### Added
+- Possibility to reorder table columns in the tables section upstairs.
+
+### Changed
+- Render internal links in text pages with no target attribute (that the Quill editor adds by default).
+- The welcome page is now a standard text page.
+- Security.groovy can now have user beans injected with `@Autowired`.
+
+### Fixed
+- Important security vulnerability that may have allowed access with forged tokens.
+- Authentication token refresh after expiration. [#430](https://github.com/ManyDesigns/Portofino/issues/430)
+- Backwards compatibility: revert `T extends Serializable` in CRUD actions, introduced in v5.2.0. [#428](https://github.com/ManyDesigns/Portofino/issues/428)
+
 ## [5.2.0] – 2020-11-30
 
 ### Added